Create QLocale's default store if missing when setDefault() is called
A check to prevent crash-on-exit if QLocale::setDefault() was called
after the default store was torn down had the side-effect of causing
setDefault() to not record data if it was called before the first
access to the default store caused it to come into being.

Change the check to test that the default store has been destroyed
and, if not, create it if it doesn't exist yet.  This refines commit
4d6572aac0eb1f75f3c810ce8e92635b956d29fc (as modified by commit
11c5c078c7743050a115a4dcc31f52caaa378e35).

Fixes: QTBUG-83016
Fixes: QTBUG-83415
Change-Id: Icbce9bd9c75d0258d403e2f90957561b5a18bdf3
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
(cherry picked from commit c7420d9cb8e15d1f094fa92cfab635a86246c670)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>