Windows: Fix QPainter warnings when using some fonts
Some fonts, like Segoe UI, have empty glyphs defined for
characters such as space. When there is a margin on the glyphs,
we would previously render them as 4x4 completely transparent
images. This is a waste of time, and in the case of DirectWrite,
we fixed it in bb67c3b68b00fb2075a971f4f0d58e1867ef7128 so that
it will always return an invalid image when the glyph does not
have dimensions.

But since we did not account for this in alphaMapBoundingBox(),
the texture glyph cache did not detect this case and would
request the empty image and try to add it to the cache.

Fixes: QTBUG-84629
Change-Id: I470a4a55577c5e9239bf77de81c2b381bf8e6a49
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>
(cherry picked from commit c2dcb7eaa19251123f9fa9546f4fe3508751b918)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>